sonar: No relationship between generic parameter and method argument
Reviewed-on: https://git.eclipse.org/r/130270
Tested-by: CI Bot
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>

+++ gdbtrace/org.eclipse.tracecompass.gdbtrace.core/src/org/eclipse/tracecompass/internal/gdbtrace/core/trace/DsfGdbAdaptor.java
@@ -782,7 +782,7 @@ public class DsfGdbAdaptor {
         // get corresponding TP data
         String tmfEventRef;
         String tmfEventSrc;
-        MIBreakpointDMData bp = fTpInfo.get(rank);
+        MIBreakpointDMData bp = fTpInfo.get((int) rank);
         if (bp != null) {
             tmfEventRef = bp.getFileName() + ":" + bp.getLineNumber() + " :: " + bp.getFunctionName(); //$NON-NLS-1$ //$NON-NLS-2$
             tmfEventSrc = bp.getFileName() + " :: " + bp.getFunctionName() + ", line: " + bp.getLineNumber(); //$NON-NLS-1$ //$NON-NLS-2$
